rename files to match class name
