+++ src/background/telemetry/multiplexing-telemetry-client.ts
